Feat/surgical procedure (#6)
* feat(surg_proc): adding SurgicalProcedure entity and relationship

* feat(surg_proc): Creating Index, Create and Edit view, with bussiness logic

* feat(surg_proc): adding Delete view

* feat(surg_proc): implementing Select2

* feat(surg_proc): implementing filter in Index view

* feat(surg_proc): retrieving all changes from rotativa branch

+++ MyPatient.Models/Patient.cs
@@ -19,7 +19,7 @@ namespace MyPatient.Models
         [Required(ErrorMessage = "El Nombre es requerido.")]
         public string Name { get; set; } = string.Empty;
         
-        public double? Weight { get; set; }
+        public float? Weight { get; set; }
 
         public int? Age { get; set; }
         
@@ -34,7 +34,7 @@ namespace MyPatient.Models
         // **********MA***********
         [Required(ErrorMessage = "Debe seleccionar un MA.")]
         [ForeignKey("MA")]
-        public int MAId { get; set; }
+        public long MAId { get; set; }
 
         public Doctor? MA { get; set; }
     }
